Fix subtraction ambiguous types, and some type hints
