+++ test/test_functions.py
@@ -803,11 +803,17 @@ class TestFunction(TestBase):
 
     def test_iff_ambiguous_type(self):
         with self.assertLogs(logger, logging.WARN) as cm:
+            # We want to assert there are no warnings, but the 'assertLogs' method does not support that.
+            # Therefore, we are adding a dummy warning, and then we will assert it is the only warning.
+            logger.warn("Dummy warning")
             self.assertEqual(
-                " | project foo = iff(boolField, time(3.0:0:0.0), foo - bar)",
+                " | project foo = iff(boolField, time(3.0:0:0.0), foo-bar)",
                 Query().project(foo=f.iff(t.boolField, timedelta(3), col.foo - col.bar)).render()
             )
-        self.assertEqual([], cm.output)
+        self.assertEqual(
+            ["WARNING:pykusto:Dummy warning"],
+            cm.output,
+        )
 
     def test_iif(self):
         # iif is just an alias to iff
